+++ test-e2e/manager-archive-device.js
@@ -1,30 +1,10 @@
 import test from 'node:test'
-import { KeyManager } from '@mapeo/crypto'
-import { MapeoManager as MapeoManagerPreMigration } from '@comapeo/core2.0.1'
-import RAM from 'random-access-memory'
-import { MapeoManager } from '../src/mapeo-manager.js'
-import Fastify from 'fastify'
 import assert from 'node:assert/strict'
 import { kIsArchiveDevice } from '../src/mapeo-project.js'
-import { ManagerCustodian } from './utils.js'
-import { fileURLToPath } from 'node:url'
-import fs from 'node:fs'
-import { temporaryDirectory } from 'tempy'
+import { createManager, ManagerCustodian } from './utils.js'
 
-const projectMigrationsFolder = new URL('../drizzle/project', import.meta.url)
-  .pathname
-const clientMigrationsFolder = new URL('../drizzle/client', import.meta.url)
-  .pathname
-
-test('Set & Get isArchiveDevice', async () => {
-  const manager = new MapeoManager({
-    rootKey: KeyManager.generateRootKey(),
-    projectMigrationsFolder,
-    clientMigrationsFolder,
-    dbFolder: ':memory:',
-    coreStorage: () => new RAM(),
-    fastify: Fastify(),
-  })
+test('Set & Get isArchiveDevice', async (t) => {
+  const manager = createManager('seed', t)
 
   // Ensure at least one project exists (tests internal setting on project)
   const projectId = await manager.createProject()
@@ -79,60 +59,4 @@ test('isArchiveDevice persists', async (t) => {
 
   assert.equal(isArchiveDevice1, false)
   assert.equal(isArchiveDevice2, isArchiveDevice1)
-})
-
-test('migration of localDeviceInfo table', async (t) => {
-  const comapeoCorePreMigrationUrl = await import.meta.resolve?.(
-    '@comapeo/core2.0.1'
-  )
-  assert(comapeoCorePreMigrationUrl, 'Could not resolve @comapeo/core2.0.1')
-  const clientMigrationsFolderPreMigration = fileURLToPath(
-    new URL('../drizzle/client', comapeoCorePreMigrationUrl)
-  )
-  const projectMigrationsFolderPreMigration = fileURLToPath(
-    new URL('../drizzle/project', comapeoCorePreMigrationUrl)
-  )
-
-  const dbFolder = temporaryDirectory()
-  const rootKey = KeyManager.generateRootKey()
-  t.after(() => fs.rm(dbFolder, { recursive: true }))
-
-  const managerPreMigration = new MapeoManagerPreMigration({
-    rootKey,
-    projectMigrationsFolder: projectMigrationsFolderPreMigration,
-    clientMigrationsFolder: clientMigrationsFolderPreMigration,
-    dbFolder,
-    coreStorage: () => new RAM(),
-    fastify: Fastify(),
-  })
-  const deviceInfo = /** @type {const} */ ({
-    name: 'Test Device',
-    deviceType: 'desktop',
-  })
-  const expectedDeviceInfo = {
-    ...deviceInfo,
-    deviceId: managerPreMigration.deviceId,
-  }
-  await managerPreMigration.setDeviceInfo(deviceInfo)
-  assert.deepEqual(
-    await managerPreMigration.getDeviceInfo(),
-    expectedDeviceInfo
-  )
-
-  // No manager.close() function yet, but should be ok
-
-  const manager = new MapeoManager({
-    rootKey,
-    projectMigrationsFolder,
-    clientMigrationsFolder,
-    dbFolder,
-    coreStorage: () => new RAM(),
-    fastify: Fastify(),
-  })
-
-  assert.deepEqual(
-    await manager.getDeviceInfo(),
-    expectedDeviceInfo,
-    'deviceInfo is migrated'
-  )
 })
